final price calculation done

+++ src/main/java/sfa/order_service/dto/response/OrderResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import sfa.order_service.constant.DiscountCoupon;
 import sfa.order_service.constant.OrderStatus;
 
 import java.util.Date;
@@ -15,9 +16,9 @@ import java.util.Date;
 public class OrderResponse {
     private Long orderId;
     private OrderStatus status;
-    private Double totalPrice;
-    private Double gstAmount;
-    private Double totalPriceWithGst;
+    private float totalPrice;
+    private float gstAmount;
+    private float totalPriceWithGst;
     private Date estimatedDeliveryDate;
 }
